+++ singlestoredb/connection.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 """SingleStoreDB connections and cursors."""
+
 import abc
 import inspect
 import io
@@ -22,6 +23,7 @@ from typing import Union
 from urllib.parse import parse_qs
 from urllib.parse import unquote_plus
 from urllib.parse import urlparse
+
 
 import sqlparams
 try:
@@ -1288,6 +1290,15 @@ class Connection(metaclass=abc.ABCMeta):
         """Access server properties managed by the SHOW statement."""
         return ShowAccessor(self)
 
+    @property
+    def vector_db(self) -> Any:
+        """
+        Get vectorstore API accessor 
+        """
+        from vectorstore import VectorDB
+        if not hasattr(self, '_vector_db'):
+            self._vector_db = VectorDB(connection=self)
+        return self._vector_db
 
 #
 # NOTE: When adding parameters to this function, you should always