+++ docs/src/whatsnew.rst
@@ -8,6 +8,11 @@ This document outlines features and improvements from each release.
    are for non-production testing and evaluation, and may include
    changes to the API.
 
+v1.14.0 - May 28, 2025
+------------------------
+* Add ``vector_db`` property to a ``Connection`` object to get the vectorstor API to the current database
+* Add ``vector_db(...)`` function to connect to a database and optain a vectorstor API 
+
 v1.13.0 - April 16, 2025
 ------------------------
 * Refactor external function server (experimental) type system